+++ src/main/java/kitchenpos/order/common/domain/OrderLineItem.java
@@ -10,7 +10,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
-import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.model.Menu;
 
 import java.math.BigDecimal;
 import java.util.UUID;
